enable title renaming for all dock_widget, removed duplicated code
