+++ qt_gui/resource/dock_widget_title_bar.ui
@@ -52,6 +52,12 @@
    </item>
    <item>
     <widget class="QLabel" name="title_label">
+     <property name="sizePolicy">
+      <sizepolicy hsizetype="Preferred" vsizetype="Preferred">
+       <horstretch>0</horstretch>
+       <verstretch>0</verstretch>
+      </sizepolicy>
+     </property>
      <property name="lineWidth">
       <number>0</number>
      </property>
@@ -61,13 +67,26 @@
     </widget>
    </item>
    <item>
+    <widget class="QLineEdit" name="title_edit">
+     <property name="sizePolicy">
+      <sizepolicy hsizetype="Expanding" vsizetype="Fixed">
+       <horstretch>1</horstretch>
+       <verstretch>0</verstretch>
+      </sizepolicy>
+     </property>
+    </widget>
+   </item>
+   <item>
     <spacer name="horizontalSpacer_2">
      <property name="orientation">
       <enum>Qt::Horizontal</enum>
      </property>
+     <property name="sizeType">
+      <enum>QSizePolicy::Expanding</enum>
+     </property>
      <property name="sizeHint" stdset="0">
       <size>
-       <width>383</width>
+       <width>8</width>
        <height>10</height>
       </size>
      </property>
